+++ source/Private/ConvertTo-TimeSpan.ps1
@@ -1,0 +1,31 @@
+<#
+    .SYNOPSIS
+        Assert that the value provided can be converted to
+
+    .PARAMETER Value
+       The time value as a string that should be converted.
+
+    .OUTPUTS
+        Returns an TimeSpan object containing the converted value, or $null if
+        conversion was not possible.
+#>
+function ConvertTo-TimeSpan
+{
+    [CmdletBinding()]
+    [OutputType([System.TimeSpan])]
+    param
+    (
+        [Parameter(Mandatory = $true, ValueFromPipeline = $true)]
+        [System.String]
+        $Value
+    )
+
+    $timeSpan = New-TimeSpan
+
+    if (-not [System.TimeSpan]::TryParse($Value, [ref] $timeSpan))
+    {
+        $timeSpan = $null
+    }
+
+    return $timeSpan
+}